backend/tests/test_decommission_endpoint.py
import requests

# Test decommission endpoint
url = "http://localhost:8000/decommission/"

try:
    response = requests.get(url)
    print(f"Status Code: {response.status_code}")
    print(f"Response: {response.text[:500]}")
except Exception as e:
    print(f"Error: {e}")
