backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
reportlab
python-multipart
email-validator
python-docx
openpyxl
qrcode
pillow
python-jose[cryptography]
passlib
bcrypt
argon2-cffi
pandas
docx2pdf
ultralytics
opencv-python-headless
